Correctly does emacs transpose.el  rotate-frame-clockwise, but with tabs inside tabgroups that ratain their order and forcus.  No problems as far as I can tell.
The trick was to put the tab group in trees placed under the main obsidian screen root node.  But the root was immutable, so needed to insert a single mutable node that had the tab group trees underneath it.  Maybe this won't cause problems with other code, as long as it correctly implements a tab group tree reversal, and just whips through the extra node without noticiing.